+++ src/assets/locales/de/gutenberg.json
@@ -1,0 +1,9 @@
+{
+  "Start typing something...": "Beschreibung...",
+  "Embed Video": "Video einbetten",
+  "Embed Livestream": "Livestream einbetten",
+  "Embed Playlist": "Playlist einbetten",
+  "linkToVideo": "Videos können Sie auf <1>udk/stream</1> hochladen",
+  "linkToStream": "Livestreams können Sie auf <1>udk/stream</1> starten",
+  "linkToPlaylist": "Playlists können Sie auf <1>udk/stream</1> erstellen"
+}
