+++ src/routes/gutenberg/blocks/heading.js
@@ -1,0 +1,45 @@
+import React from 'react'
+import { useBlockProps, RichText } from '@wordpress/block-editor'
+
+const heading = {
+  apiVersion: 2,
+  name: 'medienhaus/heading',
+  title: 'Heading',
+  category: 'text',
+  description: 'Introduce new sections and organize content to help visitors (and search engines) understand the structure of your content.',
+  keywords: ['title', 'subtitle'],
+  textdomain: 'default',
+  icon: <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path fill="none" d="M0 0H24V24H0z" /><path d="M13 20h-2v-7H4v7H2V4h2v7h7V4h2v16zm8-12v12h-2v-9.796l-2 .536V8.67L19.5 8H21z" fill="var(--color-fg)" /></svg>,
+  attributes: {
+    content: {
+      type: 'string'
+    }
+  },
+  edit: (props) => {
+    const {
+      attributes: { content },
+      setAttributes,
+      onRemove
+    } = props
+
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    const blockProps = useBlockProps()
+
+    const onChangeContent = (newContent) => {
+      setAttributes({ content: newContent })
+    }
+    return (
+      <RichText
+        {...blockProps}
+        tagName="h2"
+        placeholder="Heading"
+        onChange={onChangeContent}
+        onRemove={onRemove}
+        value={content}
+        disableLineBreaks
+      />
+    )
+  }
+}
+
+export default heading
