--- src/routes/gutenberg/blocks/mh-quote.json
@@ -1,20 +1,0 @@
-{
-    "$schema": "https://schemas.wp.org/trunk/block.json",
-    "apiVersion": 2,
-    "name": "medienhaus/quote",
-    "title": "Quote 2",
-    "category": "text",
-    "parent": [ "core/group" ],
-    "version": "1.0.0",
-    "attributes": {
-        "content": {
-            "type": "string",
-            "source": "html"
-        }
-    },
-    "example": {
-        "attributes": {
-            "content": "<p>I think, therefore I am.</p>"
-        }
-    }
-}
