+++ src/routes/gutenberg/editor.js
@@ -3,14 +3,11 @@ import {
   BlockEditorProvider,
   BlockList,
   BlockTools,
-  RichText,
   WritingFlow,
-  ObserveTyping,
-  useBlockProps
+  ObserveTyping
 } from '@wordpress/block-editor'
 import { SlotFillProvider, Popover } from '@wordpress/components'
 import { useState } from '@wordpress/element'
-import { View } from '@wordpress/primitives'
 
 import * as paragraph from '@wordpress/block-library/build/paragraph'
 import * as list from '@wordpress/block-library/build/list'
@@ -23,6 +20,11 @@ import { bold } from '@wordpress/format-library/build/bold'
 import { italic } from '@wordpress/format-library/build/italic'
 import { strikethrough } from '@wordpress/format-library/build/strikethrough'
 import { registerCoreBlocks } from '@wordpress/block-library'
+import { useTranslation } from 'react-i18next'
+import video from './blocks/video'
+import playlist from './blocks/playlist'
+import livestream from './blocks/livestream'
+import heading from './blocks/heading'
 
 // registerBlockType('medienhaus/quote', {
 //   apiVersion: 2,
@@ -224,8 +226,9 @@ import { registerCoreBlocks } from '@wordpress/block-library'
 //   }
 // })
 
-function GutenbergEditor ({ content = [], blockTypes = ['text', 'heading', 'list', 'code', 'video'], onChange }) {
+function GutenbergEditor ({ content = [], blockTypes = ['text', 'heading', 'list', 'code', 'video', 'playlist', 'livestream'], onChange }) {
   const [blocks, setBlocks] = useState(content)
+  const { t } = useTranslation('gutenberg')
 
   useEffect(() => {
     // register
@@ -242,118 +245,16 @@ function GutenbergEditor ({ content = [], blockTypes = ['text', 'heading', 'list
     if (blockTypes.includes('code')) registerCoreBlocks([code])
 
     if (blockTypes.includes('video')) {
-      registerBlockType('medienhaus/video', {
-        apiVersion: 2,
-        name: 'medienhaus/video',
-        title: 'Video',
-        category: 'text',
-        description: 'PeerTube Video',
-        keywords: ['peertube', 'video'],
-        icon: <svg viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-          <path d="M19 3H5c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h14c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm.5 16c0 .3-.2.5-.5.5H5c-.3 0-.5-.2-.5-.5V9.8l4.7-5.3H19c.3 0 .5.2.5.5v14zM10 15l5-3-5-3v6z" />
-        </svg>,
-        attributes: {
-          content: {
-            type: 'string'
-          }
-        },
-        edit: (props) => {
-          const {
-            attributes: { content },
-            setAttributes
-          } = props
-          // eslint-disable-next-line react-hooks/rules-of-hooks
-          const blockProps = useBlockProps()
-          // eslint-disable-next-line react-hooks/rules-of-hooks
-          const [urlInput, setUrlInput] = useState(content || '')
-
-          const onSubmit = (event) => {
-            if (event) event.preventDefault()
-
-            setAttributes({ content: urlInput })
-          }
-
-          const isValidUrl = (url) => {
-            return url && url.startsWith && url.startsWith('https://stream.udk-berlin.de/')
-          }
-
-          if (isValidUrl(content)) {
-            return (
-              <View {...blockProps}>
-                <iframe
-                  src={`https://stream.udk-berlin.de/videos/embed/${content.replace('https://stream.udk-berlin.de/w/', '')}`}
-                  frameBorder="0"
-                  title={content}
-                  sandbox="allow-same-origin allow-scripts"
-                  allowFullScreen="allowfullscreen"
-                  width="560"
-                  height="315"
-                  style={{ width: '100%', aspectRatio: '16 / 9', border: 'calc(var(--margin) * 0.2) solid var(--color-fg)' }}
-                />
-              </View>
-            )
-          }
-
-          return (
-            <View {...blockProps}>
-              <form onSubmit={onSubmit}>
-                <input
-                  type="url"
-                  value={urlInput}
-                  placeholder="Enter URL to embed here…"
-                  onChange={(e) => setUrlInput(e.target.value)}
-                />
-                <button type="submit" disabled={!isValidUrl(urlInput)}>Embed Video</button>
-              </form>
-            </View>
-          )
-        }
-      })
+      registerBlockType('medienhaus/video', video)
     }
-
+    if (blockTypes.includes('livestream')) {
+      registerBlockType('medienhaus/livestream', livestream)
+    }
+    if (blockTypes.includes('playlist')) {
+      registerBlockType('medienhaus/playlist', playlist)
+    }
     if (blockTypes.includes('heading')) {
-      registerBlockType('medienhaus/heading', {
-        apiVersion: 2,
-        name: 'medienhaus/heading',
-        title: 'Heading',
-        category: 'text',
-        description: 'Introduce new sections and organize content to help visitors (and search engines) understand the structure of your content.',
-        keywords: ['title', 'subtitle'],
-        textdomain: 'default',
-        icon: <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-          <path d="M6.2 5.2v13.4l5.8-4.8 5.8 4.8V5.2z" />
-        </svg>,
-        attributes: {
-          content: {
-            type: 'string'
-          }
-        },
-        edit: (props) => {
-          const {
-            attributes: { content },
-            setAttributes,
-            onRemove
-          } = props
-
-          // eslint-disable-next-line react-hooks/rules-of-hooks
-          const blockProps = useBlockProps()
-
-          const onChangeContent = (newContent) => {
-            setAttributes({ content: newContent })
-          }
-          return (
-            <RichText
-              {...blockProps}
-              tagName="h2"
-              placeholder="Heading"
-              onChange={onChangeContent}
-              onRemove={onRemove}
-              value={content}
-              disableLineBreaks
-            />
-          )
-        }
-      })
+      registerBlockType('medienhaus/heading', heading)
     }
 
     return () => {
@@ -362,6 +263,8 @@ function GutenbergEditor ({ content = [], blockTypes = ['text', 'heading', 'list
       unregisterBlockType('core/code')
       unregisterBlockType('core/list')
       unregisterBlockType('medienhaus/video')
+      unregisterBlockType('medienhaus/livestream')
+      unregisterBlockType('medienhaus/playlist')
       unregisterBlockType('medienhaus/heading')
       unregisterFormatType('core/bold')
       unregisterFormatType('core/italic')
@@ -385,7 +288,7 @@ function GutenbergEditor ({ content = [], blockTypes = ['text', 'heading', 'list
       settings={{
         canLockBlocks: false,
         codeEditingEnabled: false,
-        bodyPlaceholder: 'Placeholder',
+        bodyPlaceholder: t('Start typing something...'),
         keepCaretInsideBlock: false
       }}
       value={blocks}
